requirements.txt
certifi==2020.6.20
elasticsearch==6.8.1
numpy==1.19.1
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
urllib3==1.25.10
